Modularised the components :) checking lint next
