+++ src/components/TrackerComponents/ProgressComponent.jsx
@@ -1,68 +1,8 @@
 "use client";
 import React, { useState } from "react";
-import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
-import { Checkbox } from "@/components/ui/checkbox";
-import { ScrollArea } from "@/components/ui/scroll-area";
-import { Progress } from "@/components/ui/progress";
-import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
-import { ArrowUpCircle, Utensils, Dumbbell } from "lucide-react";
-
-const TodoItem = ({ item, onToggle }) => (
-  <div className="flex items-center px-2 py-2 space-x-2 transition-colors rounded-lg hover:bg-accent">
-    <Checkbox
-      id={item.id}
-      checked={item.completed}
-      onCheckedChange={() => onToggle(item.id)}
-    />
-    <label
-      htmlFor={item.id}
-      className={`flex-grow text-sm ${item.completed ? "line-through text-muted-foreground" : ""}`}
-    >
-      {item.text}
-    </label>
-    {item.completed && <ArrowUpCircle className="text-green-500" size={18} />}
-  </div>
-);
-
-const TodoList = ({ items, onToggle, icon: Icon }) => (
-  <Card>
-    <CardHeader>
-      <CardTitle className="flex items-center">
-        <Icon className="mr-2" />
-        {Icon === Utensils ? "Meals" : "Workouts"}
-      </CardTitle>
-    </CardHeader>
-    <CardContent>
-      <ScrollArea className="h-[300px] pr-4">
-        {items.map((item) => (
-          <TodoItem key={item.id} item={item} onToggle={onToggle} />
-        ))}
-      </ScrollArea>
-    </CardContent>
-  </Card>
-);
-
-const PerformanceSummary = ({ meals, workouts }) => {
-  const totalTasks = meals.length + workouts.length;
-  const completedTasks =
-    meals.filter((m) => m.completed).length +
-    workouts.filter((w) => w.completed).length;
-  const progressPercentage = (completedTasks / totalTasks) * 100;
-
-  return (
-    <Card className="mb-6">
-      <CardHeader>
-        <CardTitle>Today&apos;s Progress</CardTitle>
-      </CardHeader>
-      <CardContent>
-        <Progress value={progressPercentage} className="h-2 mb-2" />
-        <p className="text-sm text-muted-foreground">
-          {completedTasks} of {totalTasks} tasks completed
-        </p>
-      </CardContent>
-    </Card>
-  );
-};
+import { Utensils, Dumbbell } from "lucide-react";
+import PerformanceSummary from "@/components/cards/Summarycard";
+import TodoList from "@/components/cards/Todolistcard";
 
 const PerformancePage = () => {
   const [meals, setMeals] = useState([
@@ -75,6 +15,8 @@ const PerformancePage = () => {
     },
     { id: "m4", text: "Snack: Greek yogurt with nuts", completed: false },
   ]);
+  // use usestate and db fecthing logic here
+  // or don't .. directly call db and fetch meals here
 
   const [workouts, setWorkouts] = useState([
     { id: "w1", text: "30 minutes cardio", completed: false },
@@ -82,6 +24,7 @@ const PerformancePage = () => {
     { id: "w3", text: "15 minutes stretching", completed: false },
     { id: "w4", text: "10,000 steps", completed: false },
   ]);
+  // Same yha bhi krdo sire.
 
   const toggleTodo = (id, isMeal) => {
     const updateTodos = (todos) =>
